Refactor
There were some errors and bugs related to moving IOs from bucket to
bucket. To solve these, move to using more immutable objects and update
what methods are inherited.

Using frozensets and sets is one part of this. Also, not reusing the
pipeline for multiple runs.

The Measurer class was removed and its functionality collapsed. Also
some comments were added.

Though the chart jupyter notebook was kept, run.py is now the best way
to run the code and make sure it is up-to-date.