Add override mechanism and registry
Add new mechanic allowing functionality to be overridden. This is meant
so that we can override the wanted_move_size, latency, and
next_consumption functions of buckets.

I also added a registry for the different versions of the same
algorithm. Then they are run in a loop. This doesn't work if using with
blocking for the plot, of course.
This is missing documentation on how to use it, so, for now, I've added
a dummy second example prefetch algorithm that just returns 3 to show
how it would be used.

I also put the prefetch algorithm initial values and caps in a separate
object to make it a bit easier to keep track of and to experiment with
different configurations and algorithm combinations.