Change consumption_rate to be a function
Now you can define an override for the consumption_rate, allowing it to
change based on the current tick or the number of requests or any other
information accessible. Since the interval can change, update the
CompletedGateBucket to use different internal variables to keep track of
the next consumption accurately.

Also, add units to the representation of Rate and save the unit the user
specified the Rate in originally so that it can be displayed in the
original units.