Fix bug with running on every tick
DialBucket defaulted to having next_action() demand it run on every
tick. This makes for bad performance.

CompleteBucket needed an overhaul also. If it is empty, it doesn't need
to run until an IO has been moved into it by another bucket.

Now, if the consumption rate changes to a slower rate, it will take
until the next time at which the bucket is run according to the previous
rate for the new rate to take effect.

Wait a second this doesn't work at all, i think.